+++ src/main.rs
@@ -61,7 +61,7 @@ fn set_log_level(verbosity: Verbosity) {
 
 #[tokio::main]
 async fn main() -> ExitCode {
-    let cli = match cli::parsers::post_parse_cli(Cli::parse()) {
+    let cli = match cli::post_cli(Cli::parse()) {
         Ok(cli) => cli,
         Err(err) => {
             eprintln!("{err}");